pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "crynux-worker"
authors = [{ name = "iwehf", email = "henry.lee@crynux.ai" }]
description = "Task worker for crynux node"
readme = "README.md"
requires-python = ">=3.7"
classifiers = ["Programming Language :: Python :: 3"]
version = "0.0.1"
dependencies = [
    "sd_task",
    "gpt_task",
    "PyYAML",
    "pydantic",
    "pydantic-settings",
    "websockets",
    "tenacity",
    "whatthepatch",
    "requests",
    "packaging"
]
